Move AtomicValue resource to core.
